Update migration. Use BIGINT for asserted_location

+++ mobile_config/migrations/8_mob_rad_track_add_location.sql
@@ -1,3 +1,3 @@
 ALTER TABLE mobile_radio_tracker 
-ADD COLUMN asserted_location NUMERIC,
+ADD COLUMN asserted_location BIGINT,
 ADD COLUMN asserted_location_changed_at TIMESTAMPTZ;
